moving scripts to new folder called core_modules
